airflow/requirements.txt
numpy==1.26.4
pandas==2.2.1

scikit-learn==1.4.1.post1
catboost==1.2.3
python-dotenv==1.0.1